refactor(clawdstrike-logos): extract witness from verifier.rs
Move the witness/sample synthesis cluster (regex-HIR sampling,
path/domain/token/shell witness generators, brace-alternative consumption,
and shortest_common_supersequence) into verifier/witness.rs -- the largest
impl submodule (~760 lines, the sanctioned exception). Pure relocation:
these free fns were file-private and become pub(crate) so inheritance,
verifier_core, and the test module reach them via the root's pub(crate) use
witness glob. No logic changes.
